Refactored JavaScript
