+++ lib/sparklistener.py
@@ -29,7 +29,7 @@ class SparkListener:
                 self.reader.set_message(dat)
                 self.reader.read_message()
 
-                if self.reader.message[0][0] == 4:
+                if self.reader.message[0][0] == 4 and self.reader.message[0][1] == 0:
                     #Acknowledgement
                     continue
