+++ static/js/inputs.js
@@ -1,92 +1,3 @@
-var socket = io();
-
-// Update On/Off UI elements
-function changeOnOffState(state, effect, type) {
-  amp_preset_modified();
-  if (state === "Off") {
-    $("#" + effect + "_off").addClass("selected");
-    $("#" + effect + "_on").removeClass("selected");
-    $("#" + type + "_container").addClass("grayscale");
-  } else {
-    $("#" + effect + "_on").addClass("selected");
-    $("#" + effect + "_off").removeClass("selected");
-    $("#" + type + "_container").removeClass("grayscale");
-  }
-}
-
-// Global toasts function
-function notify_user(message) {
-  alerty.toasts(
-    message,
-    {
-      bgColor: "green",
-      fontColor: "white",
-    },
-    function () {
-      $(".loading").hide();
-    }
-  );
-}
-
-// If preset is modified, notify user by flashing the preset button
-function amp_preset_modified(){
-  $(".preset_button.selected").addClass("flash");
-}
-
-function showHideContent(effect_type, effect, visible) {
-  var chevron = $("#" + effect_type + "_showhidecontent");
-  var content = $("#" + effect + "_content");
-  if (visible) {
-    content.slideDown("slow");
-    chevron.removeClass("collapse_closed");
-    chevron.addClass("collapse_open");
-  } else {
-    content.slideUp("slow");
-    chevron.removeClass("collapse_open");
-    chevron.addClass("collapse_closed");
-    visible = false;
-  }
-}
-
-// Knob changes have to be caught by this handler explicitly stated in HTML instead of JQuery element event.
-function knobChangeEventHandler(knob) {
-  amp_preset_modified();
-  var effect = knob.dataset["id"];
-  var param = knob.dataset["parameter"];
-  var value = knob.value;
-  var data = { effect: effect, parameter: param, value: value };
-  socket.emit("change_effect_parameter", data);
-}
-
-function updateChainPreset(preset_id, name) {
-  $(".loading").show();
-
-  var data = {
-    name: name,
-    preset_id: preset_id,
-  };
-
-  $.post("/chainpreset/update", data, function (result) {
-    $("#chain_preset_container").html(result);
-    notify_user("The chain preset was saved successfully.");
-  });
-}
-
-function updatePedalPreset(effecttype, preset_id, name) {
-  $(".loading").show();
-
-  var data = {
-    effect_type: effecttype,
-    name: name,
-    preset_id: preset_id,
-  };
-
-  $.post("/pedalpreset/update", data, function (result) {
-    $("#" + effecttype + "_footer").html(result);
-    notify_user("The pedal preset was saved successfully.");
-  });
-}
-
 $(document).ready(function () {
   var start, end, diff;
   var clickTime = 500;
@@ -106,7 +17,7 @@ $(document).ready(function () {
     end = Date.now();
     diff = end - start + 1;
     if (diff > clickTime) {
-      socket.emit("store_amp_preset");
+      socket.emit("store_amp_preset");      
     } else {
       if ($(this).hasClass("selected")) {
         return;
@@ -140,7 +51,7 @@ $(document).ready(function () {
   });
 
   $(document).on("change", "[type=checkbox]", function () {
-    amp_preset_modified();
+    window.amp_preset_modified();
     var effect = $(this).data("id");
     var param = $(this).data("parameter");
     var value = 0.0;
@@ -157,7 +68,7 @@ $(document).ready(function () {
   });
 
   $(document).on("change", ".effect_selector", function () {
-    amp_preset_modified();
+    window.amp_preset_modified();
 
     var effecttype = $(this).data("type");
     var oldeffect = $(this).data("selected");
@@ -190,7 +101,7 @@ $(document).ready(function () {
       visible = false;
     }
 
-    showHideContent(effect_type, effect, visible);
+    window.showHideContent(effect_type, effect, visible);
 
     var data = { effect_type: effect_type, visible: visible };
     socket.emit("show_hide_pedal", data);
@@ -214,7 +125,7 @@ $(document).ready(function () {
             }
           );
         } else {
-          updateChainPreset(0, name);
+          window.updateChainPreset(0, name);
         }
       }
     );
@@ -225,7 +136,7 @@ $(document).ready(function () {
     if (preset_id == 0) {
       return;
     }
-    updateChainPreset(preset_id, null);
+    window.updateChainPreset(preset_id, null);
   });
 
   $(document).on("click", "#delete_chain_preset", function () {
@@ -249,7 +160,7 @@ $(document).ready(function () {
 
         $.post("/chainpreset/delete", data, function (result) {
           $("#chain_preset_container").html(result);
-          notify_user("The chain preset was deleted successfully.");
+          window.notify_user("The chain preset was deleted successfully.");
         });
       }
     );
@@ -285,7 +196,7 @@ $(document).ready(function () {
             }
           );
         } else {
-          updatePedalPreset(effect_type, 0, name);
+          window.updatePedalPreset(effect_type, 0, name);
         }
       }
     );
@@ -297,7 +208,7 @@ $(document).ready(function () {
     if (preset_id == 0) {
       return;
     }
-    updatePedalPreset(effecttype, preset_id, null);
+    window.updatePedalPreset(effecttype, preset_id, null);
   });
 
   $(document).on("click", ".delete_pedal_preset", function () {
@@ -323,7 +234,7 @@ $(document).ready(function () {
 
         $.post("/pedalpreset/delete", data, function (result) {
           $("#" + effecttype + "_footer").html(result);
-          notify_user("The pedal preset was deleted successfully.");
+          window.notify_user("The pedal preset was deleted successfully.");
         });
       }
     );
