+++ templates/layout.html
@@ -15,6 +15,7 @@
         <link rel="stylesheet" href="/css/external/alerty.css" />
     <script src="/js/external/alerty.js"></script>  
     <script src="/js/external/input-knobs.js" type="text/javascript" charset="utf-8"></script>      
+    <script src="/js/common.js" type="text/javascript" charset="utf-8"></script>
     <script src="/js/inputs.js" type="text/javascript" charset="utf-8"></script>
     <script src="/js/events.js" type="text/javascript" charset="utf-8"></script>    
     <link rel="apple-touch-icon" sizes="180x180" href="/apple-touch-icon.png">
